+++ src/json2typescript/any.js
@@ -1,0 +1,9 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var Any = (function () {
+    function Any() {
+    }
+    return Any;
+}());
+exports.Any = Any;
+//# sourceMappingURL=/Users/andreas/Documents/GIT/appvision@git.appvision.ch/unibasel/json2typescript/src/json2typescript/any.js.map