+++ src/json2typescript/any.ts
@@ -1,0 +1,4 @@
+/**
+ * Virtual any type for the property decorator.
+ */
+export class Any {}