+++ src/json2typescript/json-convert-decorators.ts
@@ -1,4 +1,5 @@
 import { MappingOptions, Settings } from "./json-convert-options";
+import { Any } from "./any";
 
 /**
  * Decorator of a class that is a custom converter.
@@ -31,14 +32,45 @@ export function JsonObject(target: any) {
  * If you decide to use a custom converter, make sure this class implements the interface JsonCustomConvert from this package.
  *
  * @param jsonPropertyName optional param (default: classPropertyName) the property name in the expected JSON object
- * @param conversionOption optional param (default: undefined), should be either the expected type (String|Boolean|Number|etc) or a custom converter class implementing JsonCustomConvert
+ * @param conversionOption optional param (default: Any), should be either the expected type (String|Boolean|Number|etc) or a custom converter class implementing JsonCustomConvert
  * @param isOptional optional param (default: false), if true, the json property does not have to be present in the object
  *
  * @returns {(target:any, key:string)=>void}
  */
-export function JsonProperty(jsonPropertyName?: string, conversionOption?: any, isOptional?: boolean): any {
+export function JsonProperty(...params: any[]): any {
 
     return function (target: any, classPropertyName: string): void {
+
+        let jsonPropertyName: string = classPropertyName;
+        let conversionOption: any = Any;
+        let isOptional: boolean = false;
+
+        switch (params.length) {
+            case 0:
+                break;
+            case 1:
+                jsonPropertyName = params[0];
+                break;
+            case 2:
+                if (params[1] === undefined) throw new Error(
+                    "Fatal error in JsonConvert. " +
+                    "It's not allowed to explicitely pass \"undefined\" as second parameter in the @JsonProperty decorator.\n\n" +
+                    "\tClass name: \n\t\t" + target.constructor.name + "\n\n" +
+                    "\tClass property: \n\t\t" + classPropertyName + "\n\n" +
+                    "Use \"Any\" to allow any type. You can import this class from \"json2typescript\".\n\n"
+                );
+                jsonPropertyName = params[0];
+                conversionOption = params[1];
+                break;
+            case 3:
+                jsonPropertyName = params[0];
+                conversionOption = params[1];
+                isOptional = params[2];
+                break;
+            default:
+                break;
+        }
+
 
         if (typeof(target[Settings.MAPPING_PROPERTY]) === "undefined") {
             target[Settings.MAPPING_PROPERTY] = [];
@@ -50,9 +82,6 @@ export function JsonProperty(jsonPropertyName?: string, conversionOption?: any, 
             jsonPropertyName = classPropertyName;
         }
 
-        if (typeof(isOptional) === "undefined") {
-            isOptional = false;
-        }
 
         let jsonPropertyMappingOptions = new MappingOptions();
         jsonPropertyMappingOptions.classPropertyName = classPropertyName;
@@ -60,7 +89,7 @@ export function JsonProperty(jsonPropertyName?: string, conversionOption?: any, 
         jsonPropertyMappingOptions.isOptional = isOptional ? isOptional : false;
 
         // Check if conversionOption is a type or a custom converter.
-        if (typeof(conversionOption) !== "undefined" && typeof(conversionOption[Settings.MAPPER_PROPERTY]) !== "undefined") {
+        if (typeof(conversionOption) !== "undefined" && conversionOption !== null && typeof(conversionOption[Settings.MAPPER_PROPERTY]) !== "undefined") {
             jsonPropertyMappingOptions.customConverter = new conversionOption();
         } else {
             jsonPropertyMappingOptions.expectedJsonType = conversionOption;